+++ com.unity.cinemachine/Runtime/Core/CameraBlendStack.cs
@@ -55,23 +55,23 @@ namespace Unity.Cinemachine
     /// <summary>
     /// Implements an overridable stack of blend states.
     /// </summary>
-    internal class CameraBlendStack : ICameraOverrideStack
+    class CameraBlendStack : ICameraOverrideStack
     {
         const float kEpsilon = UnityVectorExtensions.Epsilon;
 
-        class StackFrame : BlendSourceVirtualCamera
+        class StackFrame : NestedBlendSource
         {
             public int Id;
             public int Priority;
-            public readonly CinemachineBlend Source = new (null, null, null, 1, 0);
+            public readonly CinemachineBlend Source = new ();
             public float DeltaTimeOverride;
 
             // If blending in from a snapshot, this holds the source state
-            readonly StaticStateVirtualCamera m_Snapshot = new (CameraState.Default, string.Empty);
+            readonly SnapshotBlendSource m_Snapshot = new ();
             ICinemachineCamera m_SnapshotSource;
             float m_SnapshotBlendWeight;
 
-            public StackFrame() : base(new (null, null, null, 0, 0)) {}
+            public StackFrame() : base(new ()) {}
             public bool Active => Source.IsValid;
 
             // This is a little tricky, because we only want to take a new snapshot at the start 
@@ -81,7 +81,7 @@ namespace Unity.Cinemachine
                 if (cam == null || (cam.State.BlendHint & CameraState.BlendHints.FreezeWhenBlendingOut) == 0)
                 {
                     // No snapshot required - reset it
-                    m_SnapshotSource = null;
+                    m_Snapshot.TakeSnapshot(null);
                     return cam;
                 }
                 // A snapshot is needed
@@ -89,8 +89,7 @@ namespace Unity.Cinemachine
                 {
                     // At this point we're pretty sure this is a new blend,
                     // so we take a new snapshot of the camera state
-                    m_Snapshot.Name = cam.Name;
-                    m_Snapshot.State = cam.State;
+                    m_Snapshot.TakeSnapshot(cam);
                     m_SnapshotSource = cam;
                     m_SnapshotBlendWeight = weight;
                 }
@@ -122,18 +121,20 @@ namespace Unity.Cinemachine
 
             var frame = m_FrameStack[FindFrame(overrideId, priority)];
             frame.DeltaTimeOverride = deltaTime;
-            frame.Source.CamA = camA;
-            frame.Source.CamB = camB;
-            frame.Source.Duration = 1;
-            frame.Source.BlendCurve = s_DefaultLinearAnimationCurve;
             frame.Source.TimeInBlend = weightB;
 
-            // In case vcams are inactive game objects, make sure they get initialized properly
-            if (camA is CinemachineVirtualCameraBase vcamA)
-                vcamA.EnsureStarted();
-            if (camB is CinemachineVirtualCameraBase vcamB)
-                vcamB.EnsureStarted();
+            if (frame.Source.CamA != camA || frame.Source.CamB != camB)
+            {
+                frame.Source.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(camA, camB);
+                frame.Source.CamA = camA;
+                frame.Source.CamB = camB;
 
+                // In case vcams are inactive game objects, make sure they get initialized properly
+                if (camA is CinemachineVirtualCameraBase vcamA)
+                    vcamA.EnsureStarted();
+                if (camB is CinemachineVirtualCameraBase vcamB)
+                    vcamB.EnsureStarted();
+            }
             return overrideId;
             
             // local function to get the frame index corresponding to the ID
@@ -149,7 +150,10 @@ namespace Unity.Cinemachine
                 for (index = 1; index < count; ++index)
                     if (m_FrameStack[index].Priority > priority)
                         break;
-                m_FrameStack.Insert(index, new StackFrame { Id = withId, Priority = priority });
+                var frame = new StackFrame { Id = withId, Priority = priority };
+                frame.Source.Duration = 1;
+                frame.Source.BlendCurve = s_DefaultLinearAnimationCurve;
+                m_FrameStack.Insert(index, frame);
                 return index;
             }
         }
@@ -198,18 +202,22 @@ namespace Unity.Cinemachine
             else
             {
                 var frame = m_FrameStack[0];
-                frame.Blend.CamA = frame.Blend.CamB = null;
-                frame.Source.CamA = frame.Source.CamB = null;
+                frame.Blend.ClearBlend();
+                frame.Blend.CamB = null;
+                frame.Source.ClearBlend();
+                frame.Source.CamB = null;
             }
         }
 
         /// <summary>
         /// Call this every frame with the current active camera of the root frame.
         /// </summary>
+        /// <param name="context">The mixer context in which this blend stack exists</param>
         /// <param name="activeCamera">Current active camera (pre-override)</param>
         /// <param name="up">Current world up</param>
         /// <param name="deltaTime">How much time has elapsed, for computing blends</param>
         public void UpdateRootFrame(
+            ICinemachineMixer context,
             ICinemachineCamera activeCamera, Vector3 up, float deltaTime)
         {
             // Make sure there is a first stack frame
@@ -222,6 +230,7 @@ namespace Unity.Cinemachine
             if (activeCamera != outgoingCamera)
             {
                 bool backingOutOfBlend = false;
+                float duration = 0;
 
                 // Do we need to create a game-play blend?
                 if (LookupBlendDelegate != null && activeCamera != null && activeCamera.IsValid
@@ -236,18 +245,24 @@ namespace Unity.Cinemachine
 
                         frame.Source.CamA = outgoingCamera;
                         frame.Source.BlendCurve = blendDef.BlendCurve;
+                        duration = blendDef.BlendTime;
                     }
-                    frame.Source.Duration = blendDef.BlendTime;
+                    frame.Source.Duration = duration;
                     frame.Source.TimeInBlend = 0;
+                    frame.Source.CustomBlender = null;
                 }
                 frame.Source.CamB = activeCamera;
+
+                // Get custom blender, if any
+                if (duration > 0)
+                    frame.Source.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(outgoingCamera, activeCamera);
 
                 // Update the working blend:
                 // Check the status of the working blend.  If not blending, no problem.
                 // Otherwise, we need to do some work to chain the blends.
-                var duration = frame.Source.Duration;
+                ICinemachineCamera camA;
                 if (frame.Blend.IsComplete)
-                    frame.Blend.CamA = frame.GetSnapshotIfAppropriate(outgoingCamera, 0); // new blend
+                    camA = frame.GetSnapshotIfAppropriate(outgoingCamera, 0); // new blend
                 else
                 {
                     bool snapshot = (frame.Blend.State.BlendHint & CameraState.BlendHints.FreezeWhenBlendingOut) != 0;
@@ -259,6 +274,11 @@ namespace Unity.Cinemachine
                         && frame.Blend.Uses(activeCamera))
                         snapshot = true;
 
+                    // Avoid nesting too deeply
+                    var nbs = frame.Blend.CamA as NestedBlendSource;
+                    if (!snapshot && nbs != null && nbs.Blend.CamA is NestedBlendSource nbs2)
+                        nbs2.Blend.CamA = new SnapshotBlendSource(nbs2.Blend.CamA);
+
                     // Special case: if backing out of a blend-in-progress
                     // with the same blend in reverse, adjust the blend time
                     // to cancel out the progress made in the opposite direction
@@ -266,33 +286,62 @@ namespace Unity.Cinemachine
                     {
                         snapshot = true; // always use a snapshot for this to prevent pops
                         duration = frame.Blend.TimeInBlend;
+                        if (nbs != null)
+                            duration += nbs.Blend.Duration - nbs.Blend.TimeInBlend;
+                        else if (frame.Blend.CamA is SnapshotBlendSource sbs)
+                            duration += sbs.RemainingTimeInBlend;
                     }
 
                     // Chain to existing blend
-                    frame.Blend.CamA = snapshot 
-                        ? new StaticStateVirtualCamera(frame.Blend.State, outgoingCamera.Name)
-                        : new BlendSourceVirtualCamera(new CinemachineBlend(frame.Blend));
+                    if (snapshot)
+                        camA = new SnapshotBlendSource(frame, frame.Blend.Duration - frame.Blend.TimeInBlend);
+                    else
+                    {
+                        var blendCopy = new CinemachineBlend();
+                        blendCopy.CopyFrom(frame.Blend);
+                        camA = new NestedBlendSource(blendCopy);
+                    }
                 }
+                // For the event, we use the raw outgoing camera, not the blend source
+                frame.Blend.CamA = outgoingCamera;
                 frame.Blend.CamB = activeCamera;
                 frame.Blend.BlendCurve = frame.Source.BlendCurve;
                 frame.Blend.Duration = duration;
                 frame.Blend.TimeInBlend = 0;
+                frame.Blend.CustomBlender = frame.Source.CustomBlender;
+
+                // Allow the client to modify the blend
+                if (duration > 0)
+                    CinemachineCore.BlendCreatedEvent.Invoke(new () { Origin = context, Blend = frame.Blend });
+
+                // In case the event handler tried to change the cameras, put them back
+                frame.Blend.CamA = camA;
+                frame.Blend.CamB = activeCamera;
             }
 
             // Advance the working blend
-            if (frame.Blend.CamA != null)
-            {
-                frame.Blend.TimeInBlend += (deltaTime >= 0) ? deltaTime : frame.Blend.Duration;
-                if (frame.Blend.IsComplete)
-                {
-                    // No more blend
-                    frame.Blend.CamA = null;
-                    frame.Blend.BlendCurve = null;
-                    frame.Blend.Duration = 0;
-                    frame.Blend.TimeInBlend = 0;
-                }
-            }
+            if (AdvanceBlend(frame.Blend, deltaTime))
+                frame.Source.ClearBlend();
             frame.UpdateCameraState(up, deltaTime);
+
+            // local function
+            static bool AdvanceBlend(CinemachineBlend blend, float deltaTime)
+            {
+                bool blendCompleted = false;
+                if (blend.CamA != null)
+                {
+                    blend.TimeInBlend += (deltaTime >= 0) ? deltaTime : blend.Duration;
+                    if (blend.IsComplete)
+                    {
+                        // No more blend
+                        blend.ClearBlend();
+                        blendCompleted = true;
+                    }
+                    else if (blend.CamA is NestedBlendSource bs)
+                        AdvanceBlend(bs.Blend, deltaTime);
+                }
+                return blendCompleted;
+            }
         }
 
         /// <summary>
@@ -326,9 +375,19 @@ namespace Unity.Cinemachine
                     // Handle cases where we're blending into or out of nothing
                     if (!frame.Blend.IsComplete)
                     {
-                        frame.Blend.CamA ??= frame.GetSnapshotIfAppropriate(
-                            m_FrameStack[lastActive], frame.Blend.TimeInBlend);
-                        frame.Blend.CamB ??= m_FrameStack[lastActive];
+                        if (frame.Blend.CamA == null)
+                        {
+                            frame.Blend.CamA = frame.GetSnapshotIfAppropriate(
+                                m_FrameStack[lastActive], frame.Blend.TimeInBlend);
+                            frame.Blend.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(
+                                frame.Blend.CamA, frame.Blend.CamB);
+                        }
+                        if (frame.Blend.CamB == null)
+                        {
+                            frame.Blend.CamB = m_FrameStack[lastActive];
+                            frame.Blend.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(
+                                frame.Blend.CamA, frame.Blend.CamB);
+                        }
                     }
                     lastActive = i;
                 }
@@ -338,7 +397,8 @@ namespace Unity.Cinemachine
 
         /// <summary>
         /// Set the current root blend.  Can be used to modify the root blend state.
-        /// <param name="blend">The new blend.  If null, current blend will be force-completed.</param>
+        /// <param name="blend">The new blend.  CamA and CamB will be ignored, but the other fields 
+        /// will be taken.  If null, current blend will be force-completed.</param>
         /// </summary>
         public void SetRootBlend(CinemachineBlend blend)
         {
@@ -347,7 +407,12 @@ namespace Unity.Cinemachine
                 if (blend == null)
                     m_FrameStack[0].Blend.Duration = 0;
                 else
-                    m_FrameStack[0].Blend.CopyFrom(blend);
+                {
+                    m_FrameStack[0].Blend.BlendCurve = blend.BlendCurve;
+                    m_FrameStack[0].Blend.TimeInBlend = blend.TimeInBlend;
+                    m_FrameStack[0].Blend.Duration = blend.Duration;
+                    m_FrameStack[0].Blend.CustomBlender = blend.CustomBlender;
+                }
             }
         }
 
@@ -364,6 +429,39 @@ namespace Unity.Cinemachine
                     return frame.DeltaTimeOverride;
             }
             return -1;
+        }
+
+        /// <summary>
+        /// Static source for blending. It's not really a virtual camera, but takes
+        /// a CameraState and exposes it as a virtual camera for the purposes of blending.
+        /// </summary>
+        class SnapshotBlendSource : ICinemachineCamera
+        {
+            CameraState m_State;
+            string m_Name;
+
+            public float RemainingTimeInBlend { get; set; }
+
+            public SnapshotBlendSource(ICinemachineCamera source = null, float remainingTimeInBlend = 0)
+            {
+                TakeSnapshot(source);
+                RemainingTimeInBlend = remainingTimeInBlend;
+            }
+
+            public string Name => m_Name;
+            public string Description => Name;
+            public CameraState State => m_State;
+            public bool IsValid => true;
+            public ICinemachineMixer ParentCamera => null;
+            public void UpdateCameraState(Vector3 worldUp, float deltaTime) {}
+            public void OnCameraActivated(ICinemachineCamera.ActivationEventParams evt) {}
+
+            public void TakeSnapshot(ICinemachineCamera source)
+            {
+                m_State = source != null ? source.State : CameraState.Default; 
+                m_State.BlendHint &= ~CameraState.BlendHints.FreezeWhenBlendingOut;
+                m_Name ??= source == null ? "(null)" : "*" + source.Name;
+            }
         }
     }
 }