+++ com.unity.cinemachine/CHANGELOG.md
@@ -6,7 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+### Added
+
+- ThirdPerson Shooter sample scene now has an option to swap shoulders.
+- Added CinemachineCore.GetCustomBlender and BlendCreatedEvent to allow custom blend behaviour.
+- New Custom Blends sample scene illustrating how to customize the blend algorithm
+
 ### Changed
+- Improved handling of nested blends.
 - CinemachineChannels can now be named via the CinemachineChannelNames asset.  OutputChannel struct has been removed.